refactor(save): move PK-hash save-call suffix onto GenericDatabaseProvider
The suffix is the identity of a save call (schema.table|pk), not T-SQL
grammar. GenerateSaveSQL already lives on GenericDatabaseProvider;
SQLServerDataProvider.RenderSaveCallBinding only consumes the suffix.
PostgreSQL positional/json-arg bindings do not name locals today.
