+++ packages/SQLServerDataProvider/src/SQLServerDataProvider.ts
@@ -78,7 +78,6 @@ import { DuplicateRecordDetector } from '@memberjunction/ai-vector-dupe';
 import type { IColocatedVectorHost } from '@memberjunction/ai-vectordb';
 import type { DatabasePlatform } from '@memberjunction/sql-dialect';
 
-import { createHash } from 'crypto';
 import { v4 as uuidv4 } from 'uuid';
 import { UUIDsEqual } from '@memberjunction/global';
 import { SQLServerDialect, SQLDialect } from '@memberjunction/sql-dialect';
@@ -1108,63 +1107,9 @@ export class SQLServerDataProvider
   }
 
   /**
-   * First 8 hex of sha1(`${schema}.${table}|${pk values}`). Deterministic so
-   * MetadataSync recaptures of an unchanged tree are byte-identical. Random
-   * uuidv4 suffixes collided across large captures (loom #12 WP3 / F-D).
-   */
-  public static SaveCallVariableHash(schemaName: string, baseTable: string, pkValues: unknown[]): string {
-    const pk = pkValues.map((v) => (v === null || v === undefined ? '' : String(v))).join('|');
-    return createHash('sha1').update(`${schemaName}.${baseTable}|${pk}`).digest('hex').slice(0, 8);
-  }
-
-  /** Per transaction-group counts so the same PK twice in one batch gets `_hash_2`. */
-  private _saveCallSuffixCounts = new WeakMap<object, Map<string, number>>();
-
-  /**
-   * Variable suffix for DECLARE/SET locals in a save call.
-   *
-   * Naming contract: `_<8 lowercase hex>` from
-   * `sha1(\`${schema}.${table}|${pk values joined by |}\`)`, plus an optional
-   * `_<n>` with n ≥ 2 when that hash repeats inside one `TransactionGroup`
-   * (`_abc12345`, `_abc12345_2`, …). Downstream parsers key on this shape.
-   * MetadataSync captures without a transaction group, so `_n` does not appear
-   * in cheese migrations; WP4 of loom #12 must still widen
-   * `check-sync-id-parity.mjs` to accept `_<n>`.
-   */
-  protected allocateSaveCallSuffixForPk(
-    group: object | null | undefined,
-    schemaName: string,
-    baseTable: string,
-    pkValues: unknown[],
-  ): string {
-    const hash = SQLServerDataProvider.SaveCallVariableHash(schemaName, baseTable, pkValues);
-    if (!group) {
-      return `_${hash}`;
-    }
-    let counts = this._saveCallSuffixCounts.get(group);
-    if (!counts) {
-      counts = new Map();
-      this._saveCallSuffixCounts.set(group, counts);
-    }
-    const n = (counts.get(hash) ?? 0) + 1;
-    counts.set(hash, n);
-    return n === 1 ? `_${hash}` : `_${hash}_${n}`;
-  }
-
-  protected allocateSaveCallSuffix(entity: BaseEntity): string {
-    const pkValues = entity.PrimaryKey?.KeyValuePairs?.map((p) => p.Value) ?? [];
-    return this.allocateSaveCallSuffixForPk(
-      entity.TransactionGroup,
-      entity.EntityInfo.SchemaName,
-      entity.EntityInfo.BaseTable,
-      pkValues,
-    );
-  }
-
-  /**
    * Renders the SQL Server DECLARE/SET/EXEC binding for a save call.
-   * Emits per-field PK-hash-suffixed variables to keep batched saves
-   * (`SQLServerTransactionGroup`) collision-free. PKs on UPDATE are
+   * Suffixes come from GenericDatabaseProvider.allocateSaveCallSuffix
+   * (PK hash, not uuidv4). PKs on UPDATE are
    * tail-appended from `entity.PrimaryKey.KeyValuePairs`.
    *
    * Emits `_Clear` companion args when a nullable column carrying a
